+++ src/common/utils.ts
@@ -1,4 +1,5 @@
 import * as R from 'ramda';
+import React from 'react';
 
 export const assert = (
   expression: boolean,
@@ -20,5 +21,6 @@ export const assert = (
 };
 
 export const isUndefined = (x: any): x is undefined => typeof x === 'undefined';
+export const isElement = (x: any): x is React.ReactNode => React.isValidElement(x);
 
 export const pickAriaAttributes = R.pickBy((_, key: string) => R.test(/^aria-/, key));